+++ src/plugins/discourse/fetch.test.js
@@ -1,34 +1,10 @@
 // @flow
 
-import deepFreeze from "deep-freeze";
-import {Fetcher, type DiscourseFetchOptions} from "./fetch";
-import base64url from "base64url";
-import path from "path";
-import fs from "fs-extra";
+import {Fetcher} from "./fetch";
+import {options, snapshotFetcher} from "./mockSnapshotFetcher";
 
 describe("plugins/discourse/fetch", () => {
-  const options: DiscourseFetchOptions = deepFreeze({
-    apiKey: "FAKE_KEY",
-    apiUsername: "credbot",
-    serverUrl: "https://sourcecred-test.discourse.group",
-  });
-
   describe("snapshot testing", () => {
-    async function snapshotFetch(
-      url: string | Request | URL
-    ): Promise<Response> {
-      const snapshotDir = "src/plugins/discourse/snapshots";
-      const filename = base64url(url);
-      const file = path.join(snapshotDir, filename);
-      if (await fs.exists(file)) {
-        const contents = await fs.readFile(file);
-        return new Response(contents, {status: 200});
-      } else {
-        throw new Error(`couldn't load snapshot for ${file}`);
-      }
-    }
-    const snapshotFetcher = () => new Fetcher(options, snapshotFetch, 0);
-
     it("loads LatestTopicId from snapshot", async () => {
       const topicId = await snapshotFetcher().latestTopicId();
       expect(topicId).toMatchInlineSnapshot(`21`);
